refactor(ui_ast): add builder tests and refactor module structure
- Add comprehensive tests for UI AST builder functionality including
  composite variant titles using kind const, scalar array notation,
  and oneOf/anyOf rendering as composites
- Extract builder logic into separate modules: composite, defaults,
  key_value, naming, schema_helpers, and visit
- Move helper functions to appropriate modules for better code
  organization
- Update test module registration to include new builder tests
