+++ src/tests/ui_ast/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod builder_tests;
 pub(crate) mod defaults_tests;
 pub(crate) mod index_tests;
 pub(crate) mod layout_tests;
